Fine-tune resource usage.
The FILTERED module (`process_single`) may be CPU-bound when getting
only 1 CPU, therefore increasing it to 2.  On the other hand, it does
not need much memory, therefore reducing to 2GB.  Increasing time to 2h
as I/O seems to be a bottleneck too (unless it was only CPU).

The MultiQC modules are also using `process_single`, but they run only
once per pipeline run, so wasting 1 CPU on them is not a problem.

Assembly-scan also uses 100% CPU when given only 1, so doubling that
just in case.